# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Amafor Gladiators Digital Ecosystem API
  description: |
    Complete REST API specification with full traceability to SRS requirements.
    Each endpoint includes custom extension `x-srs-req` for requirement mapping.
  version: 1.0.0
  contact:
    name: Engineering Team
    email: dev@amaforgladiatorsfc.com
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.amaforgladiatorsfc.com/api
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /users/profile:
    get:
      tags: [Users]
      summary: Get current user profile
      operationId: API-USER-001
      x-srs-req: REQ-AUTH-02
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
    patch:
      tags: [Users]
      summary: Update user profile
      operationId: API-USER-002
      x-srs-req: REQ-AUTH-02
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Profile updated
  /users/{userId}/verify:
    patch:
      tags: [Users]
      summary: Verify user account (Admin only)
      operationId: API-USER-003
      x-srs-req: [REQ-ADM-06, REQ-ADM-11]
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User verified

  # Players
...
  /payments/advertiser:
    get:
      tags: [Payments]
      summary: Get payments by advertiser
      operationId: API-PAYMENT-001
      x-srs-req: REQ-ADV-05
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Advertiser payments
  /payments:
    get:
      tags: [Payments]
      summary: Get all payments
      operationId: API-PAYMENT-002
      x-srs-req: REQ-ADM-01
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All payments
...
  /payments/verify/{reference}:
    post:
      tags: [Payments]
      summary: Verify transaction
      operationId: API-PAYMENT-004
      x-srs-req: [REQ-SUP-01, REQ-SUP-02, REQ-ADV-04]
      security:
        - BearerAuth: []
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction verified
...
  /payments/advertisement:
    post:
      tags: [Payments]
      summary: Initiate advertisement payment
      operationId: API-PAYMENT-006
      x-srs-req: REQ-ADV-04
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Advertisement payment initiated
  /payments/donation:
    post:
      tags: [Payments]
      summary: Initiate donation payment
      operationId: API-PAYMENT-007
      x-srs-req: REQ-SUP-01
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Donation payment initiated
  /payments/{id}:
    get:
      tags: [Payments]
      summary: Get payment details
      operationId: API-PAYMENT-008
      x-srs-req: REQ-ADM-01
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment details
  /payments/user:
    get:
      tags: [Payments]
      summary: Get user payments
      operationId: API-PAYMENT-009
      x-srs-req: REQ-SUP-03
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: User payment history
  /payments/stats:
    get:
      tags: [Payments]
      summary: Get payment statistics
      operationId: API-PAYMENT-010
      x-srs-req: REQ-ADM-01
      security:
        - BearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Payment statistics
  /payments/{id}/refund:
    post:
      tags: [Payments]
      summary: Refund payment
      operationId: API-PAYMENT-011
      x-srs-req: REQ-ADM-01
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment refunded

  # Academy Staff
...